fix(publish): restore dist/ npm layout for CDN and bundler compatibility
Publish from package roots again so dist/bundle.js matches legacy jsdelivr URLs.
Consolidate build scripts, commit per-package LICENSE files, and drop dist/package.json copying.
